switched all code to console instead of menu
